+++ NEWS.md
@@ -1,5 +1,7 @@
 # crunch 1.27.5 (Development version)
 * Renaming variable folders via `setName()` is fixed (#487)
+* You can set the weight and filter to the same value for every slide in a deck by assigning a `weight` or `filter` on the deck object.
+* You can now derive a variable directly from a logical expression (#286)
 
 # crunch 1.27.4
 * Fixes to pkgdown documentation website